KERNELCI: x86_64_defconfig: Enable support for Chromebooks devices
As part of KernelCI we plan to add different x86 based Chromebooks to do
test boot and runtime testing. It will be useful have an upstream
defconfig supporting this kind of devices like we have for ARM
(multi_v7_defconfig) and ARM64 (defconfig). So add different options to
enable different Chromebooks.

Signed-off-by: Enric Balletbo i Serra <enric.balletbo@collabora.com>
Signed-off-by: Benson Leung <bleung@chromium.org>